[#36]refactor: 파일 구조 리팩토링 및 컴포넌트 정리
- components/Layout/VoteNav -> components/Layout/Vote로 구조 변경
  - VotingModal 및 관련 파일 이동 및 정리
- hooks/vote 디렉터리 추가 및 투표 관련 훅 분리
  - useProcessProgress, useVoteAnimation, useVoteProgress 추가
- store/vote 디렉터리 추가 및 상태 관리 관련 파일 분리
  - colorStore, processStore, voteStore 추가
- Puzzle, StageGimmicks, Templates 등 기타 관련 파일 재구조화
- 코드 가독성 및 유지보수성을 위한 디렉터리 구조 개선
